+++ test/collective-robot-behaviour-test/communication_test.cc
@@ -14,18 +14,6 @@ namespace centralised_ai
 {
 namespace collective_robot_behaviour
 {
-
-TEST(ComputeRewardsTest, TestShape)
-{
-    torch::Tensor states = torch::ones(43);
-    RewardConfiguration reward_configuration = {1, 1, 1};
-    Team own_team = Team::kBlue;
-
-    torch::Tensor output = ComputeRewards(states, reward_configuration, own_team);
-
-    EXPECT_EQ(output.size(0), 6);
-}
-
 /* Mock class for VisionClient. */
 class VisionClientDerived : public centralised_ai::ssl_interface::VisionClient
 {
@@ -69,22 +57,19 @@ static void SetAllPositionsToZero(VisionClientDerived& vision_client)
   vision_client.SetBallPositionY(0.0F);
 }
 
-/**/
-//TEST(GetStateTest, TestShape)
-//{
-//    VisionClientDerived vision_client = VisionClientDerived("127.0.0.1", 20001);
-//    centralised_ai::ssl_interface::AutomatedReferee automated_referee(vision_client, "127.0.0.1", 10001);
 
-//   SetAllPositionsToZero(vision_client);
+TEST(GetGlobalStateTest, TestShape)
+{
+  VisionClientDerived vision_client = VisionClientDerived("127.0.0.1", 20001);
+  centralised_ai::ssl_interface::AutomatedReferee automated_referee(vision_client, "127.0.0.1", 10001);
 
-//    RewardConfiguration reward_configuration = {1, 1, 1};
-
-//    torch::Tensor states = GetStates(automated_referee, vision_client, Team::kBlue, Team::kYellow);
-
-//    EXPECT_EQ(states.size(0), 1);
-//    EXPECT_EQ(states.size(1), 1);
-//    EXPECT_EQ(states.size(2), 43);
-//}
+  SetAllPositionsToZero(vision_client);
+  RewardConfiguration reward_configuration = {1, 1, 1, 1};
+  torch::Tensor state = GetGlobalState(automated_referee, vision_client, Team::kBlue, Team::kYellow);
+  EXPECT_EQ(state.size(0), 1);
+  EXPECT_EQ(state.size(1), 1);
+  EXPECT_EQ(state.size(2), 42);
+}
 
 
 TEST(ComputeOpponentTeamTest, TestOpponentTeam_1)
@@ -104,5 +89,6 @@ TEST(ComputeOpponentTeamTest, TestOpponentTeam_2)
     EXPECT_EQ(own_team, Team::kYellow);
     EXPECT_EQ(opponent_team, Team::kBlue);
 }
+
 }
 }