+++ docs/manual/changelog.md
@@ -2,6 +2,10 @@
 
 @page changelog Changelog
 
+2024-11-26
+-----------------------
+- Moved Simulation interface to this repo
+
 2024-10-29
 -----------------------
 - Added changelog, developer_guide and user_guide pages for doxygen.